+++ Almanac/UI/NPCCustomization.cs
@@ -60,10 +60,8 @@ public class NPCCustomization : MonoBehaviour, IDragHandler, IBeginDragHandler, 
         
         transform.position = Configs.CustomizationPos;
         View.OnScrollbarSensitivityChanged(Configs.ScrollbarSensitivity);
-    }
-
-    public void Start()
-    {
+        
+        
         background.SetBackground(Configs.bkgOption);
 
         mainButton.onClick.AddListener(OnMainButton);
@@ -129,7 +127,77 @@ public class NPCCustomization : MonoBehaviour, IDragHandler, IBeginDragHandler, 
         SetMainButton("Save");
         SetTopic("NPC Customization");
         Hide();
+        
     }
+
+    // public void Start()
+    // {
+    //     background.SetBackground(Configs.bkgOption);
+    //
+    //     mainButton.onClick.AddListener(OnMainButton);
+    //
+    //     npcName = new NPCSetting(SettingType.Name);
+    //     npcName.SetDescription("Name of the NPC");
+    //     npcName.SetTitle(Keys.Name);
+    //     dialogueID = new NPCSetting(SettingType.Dialogue);
+    //     dialogueID.SetDescription("Entry Dialogue ID");
+    //     dialogueID.SetTitle(Keys.EntryDialogueID);
+    //     randomTalk = new NPCSetting(SettingType.RandomTalk);
+    //     randomTalk.SetDescription("Random Talk ID");
+    //     randomTalk.SetTitle(Keys.RandomTalkID);
+    //     helm = new NPCSetting(SettingType.Helmet);
+    //     helm.SetTitle(Keys.Helmet);
+    //     helm.SetDescription("Helmet ItemID, ex: HelmetIron");
+    //     chest = new NPCSetting(SettingType.Chest);
+    //     chest.SetDescription("Chest ItemID, ex: ArmorIronChest");
+    //     chest.SetTitle(Keys.ChestItem);
+    //     legs = new NPCSetting(SettingType.Legs);
+    //     legs.SetDescription("Legs ItemID, ex: ArmorIronLegs");
+    //     legs.SetTitle(Keys.LegItem);
+    //     right = new NPCSetting(SettingType.RightHand);
+    //     right.SetDescription("Right Hand ItemID, ex: SwordIron");
+    //     right.SetTitle(Keys.RightHandItem);
+    //     left = new NPCSetting(SettingType.LeftHand);
+    //     left.SetDescription("Left Hand ItemID, ex: KnifeChitin");
+    //     left.SetTitle(Keys.LeftHandItem);
+    //     shoulder =  new NPCSetting(SettingType.Shoulder);
+    //     shoulder.SetDescription("Shoulder ItemID, ex: CapeLinen");
+    //     shoulder.SetTitle(Keys.ShoulderItem);
+    //     util = new NPCSetting(SettingType.Utility);
+    //     util.SetDescription("Utility ItemID, ex: BeltStrength");
+    //     util.SetTitle(Keys.UtilityItem);
+    //     backRight = new NPCSetting(SettingType.BackRight);
+    //     backRight.SetDescription("Back Right ItemID, ex: ShieldWood");
+    //     backRight.SetTitle(Keys.BackRightItem);
+    //     backLeft = new NPCSetting(SettingType.BackLeft);
+    //     backLeft.SetDescription("Back Left ItemID, ex: AtgeirIron");
+    //     backLeft.SetTitle(Keys.BackLeftItem);
+    //     hair = new NPCSetting(SettingType.Hair);
+    //     hair.SetDescription("Hair ItemID, ex: Hair20");
+    //     hair.SetTitle(Keys.HairItem);
+    //     beard = new NPCSetting(SettingType.Beard);
+    //     beard.SetDescription("Beard ItemID, ex: Beard3");
+    //     beard.SetTitle(Keys.BeardItem);
+    //     anim = new NPCSetting(SettingType.Animation);
+    //     anim.SetDescription("Animation ID, ex: work, stir, forge, dance");
+    //     anim.SetTitle(Keys.Animation);
+    //     skin = new NPCSetting(SettingType.SkinColor);
+    //     skin.SetDescription("Skin Color ex: 0.9 0.8 0.7");
+    //     skin.SetTitle(Keys.SkinColor);
+    //     hairColor = new NPCSetting(SettingType.HairColor);
+    //     hairColor.SetDescription("Hair/Beard Color ex: 1 0.5 0");
+    //     hairColor.SetTitle(Keys.HairColor);
+    //     modelIndex = new NPCSetting(SettingType.ModelIndex, InputField.ContentType.DecimalNumber);
+    //     modelIndex.SetDescription("Model Index, ex: 0 = male, 1 = female");
+    //     modelIndex.SetTitle(Keys.ModelIndex);
+    //     scale = new NPCSetting(SettingType.Scale);
+    //     scale.SetDescription("Scale, ex: 1.5, 1.5 1.5");
+    //     scale.SetTitle("Scale");
+    //     view?.Resize(view.Count);
+    //     SetMainButton("Save");
+    //     SetTopic("NPC Customization");
+    //     Hide();
+    // }
     public void Update()
     {
         if (!IsVisible()) return;